Create a wrapper component for login/sign up pages and create components from login/sign up forms
